Add new image assets, models, and update routes for processing

+++ database/migrations/2025_02_25_084135_create_districts_table.php
@@ -16,8 +16,8 @@ class CreateDistrictsTable extends Migration
         Schema::create('districts', function (Blueprint $table) {
             $table->string("code")->primary();
             $table->string("name");
-            $table->integer("longitude");
-            $table->integer("latitude");
+            $table->float("longitude");
+            $table->float("latitude");
             $table->timestamps();
         });
     }
